audio profiles no logic re-org
